requirements.txt
Flask
Flask_SocketIO
docker-py
eventlet
requests